Split Uavcan config sending class in dedicated module
